Struct metrics_controller::controller::MetricsController

pub struct MetricsController {
    // some fields omitted
}

The metrics controller for the CD Metrics Library

Methods

impl MetricsController

fn new(is_active: bool, app_name: String, app_version: String, app_update_channel: String, app_build_id: String, app_platform: String, locale: String, device: String, arch: String) -> MetricsController

Constructs a new MetricsController. Caller passes information about their application and environment and also whether the controller should be active (should be inactive, for example, if the user has opted-out of metrics collection).

Examples

use metrics_controller::controller::MetricsController;
let mc = MetricsController::new(
    true,
    "foxbox".to_string(),
    "1.0".to_string(),
    "beta".to_string(),
    "20160522".to_string(),
    "rust".to_string(),
    "en-us".to_string(),
    "RPi2".to_string(),
    "arm".to_string());

fn send_crash_ping(self, meta_data: String) -> bool

Sends crash details to the Mozilla telemetry server. Details include the environment information specified when instantiating the MetricsController as well as metadata regarding the crash that is being reported (for example, the call stack). The 'meta_data' parameter is a JSON string.

Examples

   mc.send_crash_ping("{
       \"metadata\": {
           \"callstack\": \"exception in thread 'main' NullPointerException ...\"
       }
   }".to_string());